add some missing casts
git-svn-id: https://svn.r-project.org/R/trunk@76293 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -332,7 +332,7 @@ static double buff_seek(Rconnection con, double where, int origin, int rw)
     { /* current */
         if (where < unread_len)
         {
-            con->buff_pos += where;
+            con->buff_pos += (size_t)where;
             return con->seek(con, NA_REAL, origin, rw);
         }
         else
